[*] 更新前端 [+] 登陆按钮开关 [+] 指定账号解析(管理员) [+] 批量复制邀请码 [+] 支持批量复制不包括格式 [*] 优化 `motrix` 按钮点击打开新页面 [*] 修复禁止删除自带分组 [*] 修复获取删除的用户分组 [+] 卡密绑定ip [*] 修复无法修改卡密到期时间

+++ app/Http/Controllers/GroupController.php
@@ -49,7 +49,7 @@ class GroupController extends Controller
         $group = Group::query()->find($group_id);
         if (!$group) return ResponseController::groupNotExists();
 
-        if (in_array($group_id, [1, 2])) {
+        if ($group_id === "1") {
             $request["name"] = $group["name"];
         } else {
             $Group = Group::query()->firstWhere("name", $request["name"]);
@@ -74,7 +74,7 @@ class GroupController extends Controller
 
         if ($validator->fails()) return ResponseController::paramsError();
 
-        if (in_array([1, 2], $request["group_ids"])) return ResponseController::groupCanNotBeRemoved("自带分组禁止删除");
+        if (in_array(1, $request["group_ids"]) || in_array(2, $request["group_ids"])) return ResponseController::groupCanNotBeRemoved("自带分组禁止删除");
 
         Group::query()->whereIn("id", $request["group_ids"])->delete();
